refactor: split shell context, side menu, dialogue, and production row parts (#3878)
Extract library parts for shell player context provider, game side menu panel/scrim,
tribe first-contact overlay flow/build, production allocation row slider/controls,
and main-menu footer quit chip to continue colonizethis_app modularization.

Refs #3878
