+++ app/lib/widgets/main_menu_footer_quit_button.dart
@@ -1,0 +1,123 @@
+part of 'main_menu.dart';
+
+/// Secondary footer Quit chip for the `pixelArt` variant. Implements
+/// `SPEC/ui/main-menu.md` § Variant rendering — Quit button row and AC 9
+/// from issue #2860: smaller than the primary wood-panel buttons, `--muted`
+/// foreground, border-only chrome (`--border` top/bottom and 1px left/right
+/// edges), and no brass corner brackets. Mirrors the mockup `.quit-btn`
+/// element in `SPEC/ui/mockups/SHEL10002-main-menu.html`.
+class _FooterQuitButton extends StatefulWidget {
+  const _FooterQuitButton({required this.label, required this.onPressed});
+
+  final String label;
+  final VoidCallback onPressed;
+
+  @override
+  State<_FooterQuitButton> createState() => _FooterQuitButtonState();
+}
+
+class _FooterQuitButtonState extends State<_FooterQuitButton> {
+  bool _hovered = false;
+
+  void _setHover(bool entered) {
+    if (_hovered == entered) return;
+    setState(() => _hovered = entered);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    final TextStyle baseStyle =
+        theme.textTheme.labelLarge ??
+        theme.textTheme.bodyMedium ??
+        const TextStyle();
+    final Color foreground = _hovered
+        ? EditorialMonoclePalette.accentBright
+        : EditorialMonoclePalette.muted;
+    final LinearGradient gradient = _hovered
+        ? LinearGradient(
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: <Color>[
+              EditorialMonoclePalette.surfaceLite,
+              EditorialMonoclePalette.surface,
+            ],
+          )
+        : LinearGradient(
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: <Color>[
+              EditorialMonoclePalette.surface,
+              EditorialMonoclePalette.bgDeep,
+            ],
+          );
+
+    return MouseRegion(
+      onEnter: (_) => _setHover(true),
+      onExit: (_) => _setHover(false),
+      cursor: SystemMouseCursors.click,
+      child: Semantics(
+        button: true,
+        label: widget.label,
+        child: GestureDetector(
+          behavior: HitTestBehavior.opaque,
+          onTap: widget.onPressed,
+          child: ConstrainedBox(
+            key: const Key(kMainMenuFooterQuitKey),
+            constraints: const BoxConstraints(
+              minHeight: kMainMenuFooterQuitMinHeight,
+              minWidth: 120,
+              maxWidth: kMainMenuFooterQuitMaxWidth,
+            ),
+            child: DecoratedBox(
+              decoration: BoxDecoration(
+                gradient: gradient,
+                border: Border(
+                  top: BorderSide(
+                    color: EditorialMonoclePalette.border,
+                    width: 1,
+                  ),
+                  bottom: BorderSide(
+                    color: EditorialMonoclePalette.border,
+                    width: 1,
+                  ),
+                  left: BorderSide(
+                    color: EditorialMonoclePalette.border,
+                    width: 1,
+                  ),
+                  right: BorderSide(
+                    color: EditorialMonoclePalette.border,
+                    width: 1,
+                  ),
+                ),
+              ),
+              child: Padding(
+                padding: const EdgeInsets.symmetric(
+                  horizontal: CtSpacing.xl,
+                  vertical: CtSpacing.m,
+                ),
+                child: Center(
+                  child: Text(
+                    widget.label,
+                    style: baseStyle.copyWith(
+                      color: foreground,
+                      fontFamily: editorialMonocleDisplayFontFamily,
+                      fontSize: kMainMenuFooterQuitFontSize,
+                      // Kept at 1.4 (distinct from the wood-panel button
+                      // letter-spacing constants) so the narrow/wide button
+                      // letter-spacing ACs do not pick up the Quit chip
+                      // label; ≈0.1em at the 12 dp chip font per mockup
+                      // `.quit-btn { letter-spacing: 0.1em }`.
+                      letterSpacing: 1.4,
+                      fontWeight: FontWeight.w600,
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
